Move format module to a separate folder
