+++ src/components/MatchDetails/MatchDetails.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { getFormattedDatetime } from '../../format';
+import { getFormattedDatetime } from '../../format/format';
 import MatchTeamScoreboard from './MatchTeamScoreboard/MatchTeamScoreboard';
 
 const MatchDetails = props => (
